Feature tela sign up perfil

+++ database/migrations/2023_05_12_150720_create_members_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('members', function (Blueprint $table) {
             $table->integer('id_members', true);
-            $table->integer('id_user')->index('members_ibfk_2');
+            $table->unsignedBigInteger('id_user')->index('members_ibfk_2');
             $table->integer('id_project')->index('members_ibfk_1');
             $table->integer('level')->index('members_ibfk_3');
         });
